Add day01 solution
Golang took some getting used to; I have to remember how to use loops
again vs. more functional constructs.

I think the heap-based approach gives us an elegant interface to our
main data structure. But in the end, the performance is exactly the same
as if we'd just used an array of three items- under the hood that's all
the heap is (with this implementation).